+++ src/radical/pilot/controller/pilot_manager_controller.py
@@ -181,7 +181,7 @@ class PilotManagerController(threading.Thread):
         """
         for worker in self._pilot_launcher_worker_pool:
             logger.debug("pworker %s disables launcher %s" % (self.name, worker.name))
-            worker.disable ()
+            worker.disable()
             logger.debug("pworker %s disabled launcher %s" % (self.name, worker.name))
 
 
@@ -225,18 +225,27 @@ class PilotManagerController(threading.Thread):
                 {'timestamp' : timestamp(), 
                  'state'     : new_state})
 
-        for [cb, cb_data] in self._shared_data[pilot_id]['callbacks']:
-            try:
-                if  self._shared_data[pilot_id]['facade_object'] :
-                    if  cb_data :
-                        cb (self._shared_data[pilot_id]['facade_object'](), new_state, cb_data)
+        try:
+            for cb in self._shared_data[pilot_id]['callbacks']:
+                cb_func = cb['cb_func']
+                cb_data = cb['cb_data']
+                try:
+                    if self._shared_data[pilot_id]['facade_object'] :
+                        if cb_data:
+                            cb_func(self._shared_data[pilot_id]['facade_object'](), new_state, cb_data)
+                        else:
+                            cb_func(self._shared_data[pilot_id]['facade_object'](), new_state)
                     else :
-                        cb (self._shared_data[pilot_id]['facade_object'](), new_state)
-                else :
-                    logger.error("Couldn't call callback (no pilot instance)")
-            except Exception as e:
-                logger.exception("Couldn't call callback function %s" % e)
-                raise
+                        logger.error("Couldn't call callback (no pilot instance)")
+                except Exception as e:
+                    logger.exception("Couldn't call callback function %s" % e)
+                    raise
+        except SystemExit:
+            # one of the callbacks requested a sys exit.  We don't want the
+            # callbacks to get into te way of the shutdown, so we unregister
+            # them all right here
+            self.unregister_pilot_callback(pilot_id)
+            raise
 
         # If we have any manager-level callbacks registered, we
         # call those as well!
@@ -250,15 +259,13 @@ class PilotManagerController(threading.Thread):
                 else :
                     logger.error("Couldn't call manager callback (no pilot instance)")
             except Exception as e:
-                logger.exception(
-                    "Couldn't call callback function %s" % e)
+                logger.exception("Couldn't call callback function %s" % e)
                 raise
 
         # if we meet a final state, we record the object's callback history for
         # later evalutation
-        if  new_state in (DONE, FAILED, CANCELED) :
+        if new_state in (DONE, FAILED, CANCELED):
             self._dbs.publish_compute_pilot_callback_history (pilot_id, self._callback_histories[pilot_id])
-      # print 'publishing Callback history for %s' % pilot_id
 
 
     # ------------------------------------------------------------------------
@@ -455,11 +462,12 @@ class PilotManagerController(threading.Thread):
 
     # ------------------------------------------------------------------------
     #
-    def register_pilot_callback(self, pilot, callback_func, callback_data=None):
+    def register_pilot_callback(self, pilot, cb_func, cb_data=None):
         """Registers a callback function.
         """
         pilot_uid = pilot.uid
-        self._shared_data[pilot_uid]['callbacks'].append([callback_func, callback_data])
+        self._shared_data[pilot_uid]['callbacks'].append({'cb_func' : cb_func, 
+                                                          'cb_data' : cb_data})
 
         # Add the facade object if missing, e.g., after a re-connect.
         if  self._shared_data[pilot_uid]['facade_object'] is None:
@@ -475,10 +483,29 @@ class PilotManagerController(threading.Thread):
 
     # ------------------------------------------------------------------------
     #
-    def register_manager_callback(self, callback_func, callback_data=None):
+    def unregister_pilot_callback(self, pid, cb_func=None):
+        """
+        Un-registers a callback function -- or all if cb_func is None
+        """
+        if not pid in self._shared_data:
+            raise ValueError('unknown pilot %s' % pid)
+
+        if cb_func:
+            # iterate over copy
+            for cb in self._shared_data[pid]['callbacks'][:]: 
+                if cb_func == cb['cb_func']:
+                    self._shared_data[pid]['callbacks'].remove(db)
+        else:
+            # remove all callbacks
+            self._shared_data[pid]['callbacks'] = []
+
+
+    # ------------------------------------------------------------------------
+    #
+    def register_manager_callback(self, cb_func, cb_data=None):
         """Registers a manager-level callback.
         """
-        self._manager_callbacks.append([callback_func, callback_data])
+        self._manager_callbacks.append([cb_func, cb_data])
 
     # ------------------------------------------------------------------------
     #
@@ -504,56 +531,88 @@ class PilotManagerController(threading.Thread):
         # not do that, it will get killed the hard way...
         delayed_cancel = list()
 
-        for pilot_id in pilot_ids :
-            if  pilot_id in self._shared_data :
+        for pilot_id in pilot_ids:
+            if  pilot_id in self._shared_data:
 
-                # read state fomr _shared_data only once, so that it does not
+                # we don't want to see any callbacks during shutdown
+                # FXIME: this is semantically incorrect, as a CANCELED state cb
+                #        should still be issued.  But, shutdown races screw us
+                #        once more...
+                self.unregister_pilot_callback(pilot_id)
+
+                # read state from _shared_data only once, so that it does not
                 # change under us...
                 old_state = str(self._shared_data[pilot_id]["data"]["state"])
 
-                logger.warn ("actively cancel pilot %s state: %s" % (pilot_id, old_state))
-                if  old_state in [DONE, FAILED, CANCELED] :
-                    logger.warn ("can't actively cancel pilot %s: already in final state" % pilot_id)
+                if old_state in [DONE, FAILED, CANCELED]:
+                    logger.debug("can't actively cancel pilot %s: already in final state" % pilot_id)
 
-                elif old_state in [PENDING_LAUNCH, LAUNCHING, PENDING_ACTIVE] :
-                    if pilot_id in self._shared_worker_data['job_ids'] :
+                elif old_state in [PENDING_LAUNCH, LAUNCHING, PENDING_ACTIVE]:
 
-                        try :
+                    if pilot_id in self._shared_worker_data['job_ids']:
+
+                        try:
                             job_id, js_url = self._shared_worker_data['job_ids'][pilot_id]
                             self._shared_data[pilot_id]["data"]["state"] = CANCELING
-                            logger.info ("actively cancel pilot %s (%s, %s)" % (pilot_id, job_id, js_url))
+                            logger.info("actively cancel pilot %s (%s, %s)" % (pilot_id, job_id, js_url))
 
                             js = self._shared_worker_data['job_services'][js_url]
-                            job = js.get_job (job_id)
-                            job.cancel ()
-                        except Exception as e :
-                            logger.exception ('pilot cancelation failed')
+                            job = js.get_job(job_id)
+                            job.cancel()
+                        except Exception as e:
+                            logger.exception('pilot cancelation failed')
 
 
-                    else :
-                        logger.warn ("can't actively cancel pilot %s: no job id known" % pilot_id)
-                        logger.debug (pprint.pformat (self._shared_worker_data))
+                    else:
+                        logger.debug("can't (yet) cancel starting pilot %s" % pilot_id)
+                        delayed_cancel.append(pilot_id)
 
-                else :
-                    logger.debug ("delay to actively cancel pilot %s: state %s" % (pilot_id, old_state))
-                    delayed_cancel.append (pilot_id)
+                else:
+                    logger.debug("delay to actively cancel pilot %s: state %s" % (pilot_id, old_state))
+                    delayed_cancel.append(pilot_id)
 
-            else :
-                logger.warn  ("can't actively cancel pilot %s: unknown pilot" % pilot_id)
-                logger.debug (pprint.pformat (self._shared_data))
+            else:
+                raise RuntimeError("unknown pilot" % pilot_id)
 
         # now tend to all delayed cancellation requests (ie. active pilots) --
         # if there are any
         if  delayed_cancel :
 
-            # grant some levay to the unruly children...
-            time.sleep (10)
+            # grant some levay (30 sec) to the unruly children...
+            delay = 30
+            start = time.time()
+            logger.report.idle(mode='start')
+            logger.report.idle(c=' ')
 
+            # we idle as long as any pilot is left to watch and we did not yet
+            # hit the delay timeout
+            while delayed_cancel and time.time() - start < delay:
+
+                logger.report.idle()
+                for pilot_id in delayed_cancel[:]:
+
+                    if pilot_id not in self._shared_data:
+                        continue
+
+                    state = self._shared_data[pilot_id]["data"]["state"]
+                    if state in [DONE, FAILED, CANCELED]:
+                        delayed_cancel.remove(pilot_id)
+
+                time.sleep(0.3)
+            logger.report.idle(mode='stop')
+
+            # force-kill any remaining pilots
             for pilot_id in delayed_cancel :
 
                 if pilot_id in self._shared_worker_data['job_ids'] :
 
                     try :
+                        # FIXME: the SAGA layer may raise an exception, because
+                        # there is a race between the delay state check above
+                        # and the cancel below, and the pilot may be gone.  That
+                        # is not a problem, we can ignore it -- what is
+                        # a problem is that the SAGA logger will complain rather
+                        # loundly...
                         job_id, js_url = self._shared_worker_data['job_ids'][pilot_id]
                         logger.info ("actively cancel pilot %s (delayed) (%s, %s)" % (pilot_id, job_id, js_url))
 
@@ -561,11 +620,11 @@ class PilotManagerController(threading.Thread):
                         job = js.get_job (job_id)
                         job.cancel ()
                     except Exception as e :
-                        logger.info ('delayed pilot cancelation failed. '
-                                     'This is not necessarily a problem.')
+                        logger.info('delayed pilot cancelation failed. '
+                                    'This is not necessarily a problem.')
 
                 else :
-                    logger.warn ("can't actively cancel pilot %s: no job id known (delayed)" % pilot_id)
+                    logger.error("can't actively cancel pilot %s: no job id known (delayed)" % pilot_id)
                     logger.debug (pprint.pformat (self._shared_worker_data))