+++ src/radical/pilot/scheduler/backfilling.py
@@ -152,13 +152,15 @@ class BackfillingScheduler(Scheduler):
                       #     logger.debug ('unit %s frees %s cores on (-> %s)' \
                       #                % (uid, unit.description.cores, pid, self.pilots[pid]['caps']))
 
-                    if not found_unit :
-                        logger.warn ('unit %s freed %s cores on %s (== %s) -- not reused'
-                                  % (uid, unit.description.cores, pid, self.pilots[pid]['caps']))
+                  # FIXME: this warning should not come up as frequently as it
+                  #        does -- needs investigation!
+                  # if not found_unit :
+                  #     logger.warn ('unit %s freed %s cores on %s (== %s) -- not reused'
+                  #               % (uid, unit.description.cores, pid, self.pilots[pid]['caps']))
 
 
         except Exception as e :
-            logger.error ("error in unit callback for backfiller (%s) - ignored" % e)
+            logger.exception ("error in unit callback for backfiller (%s) - ignored" % e)
 
 
     # -------------------------------------------------------------------------
@@ -434,7 +436,8 @@ class BackfillingScheduler(Scheduler):
 
                 # sanity check on unit state
                 if  unit.state not in [NEW, SCHEDULING, UNSCHEDULED] :
-                    raise RuntimeError ("scheduler queue should only contain NEW or UNSCHEDULED units (%s)" % uid)
+                    raise RuntimeError ("scheduler requires NEW or UNSCHEDULED units (%s:%s)"\
+                                    % (uid, unit.state))
 
               # logger.debug ("examine unit  %s (%s cores)" % (uid, ud.cores))