+++ src/radical/pilot/utils/component.py
@@ -329,7 +329,13 @@ class Component(mp.Process):
                 self.finalize()
                 self._prof.prof("finalized")
                 self._prof.flush()
+            # Signal the child
+            self._log.debug('Signalling child')
             self.terminate()
+            # Wait for the child process
+            self._log.debug('Waiting for child')
+            self.join()
+            self._log.debug('Child done')
 
         else:
             if not self._finalized:
@@ -747,6 +753,8 @@ class Component(mp.Process):
             self._log.exception('loop exception')
 
         except:
+            # This is most likely a sys.exit, but can be any other signal or
+            # interrupt.
             self._log.exception('loop error')
 
         finally:
@@ -786,6 +794,7 @@ class Component(mp.Process):
                 # send state notifications
                 self.publish('state', unit)
                 self._prof.prof('publish', uid=unit['_id'], state=unit['state'])
+                time.sleep(0.1)
 
             if push:
                 if state not in self._outputs: